+++ packages/design-system-twrnc-preset/src/index.ts
@@ -3,7 +3,10 @@ export { ThemeProvider } from './ThemeProvider';
 export { Theme } from './Theme.types';
 
 // Hooks
-export { useTailwind, useTheme } from './hooks';
+export { useTailwind, useTheme, usePureBlack } from './hooks';
+
+// Theme colors
+export { getThemeColors } from './colors';
 
 // Config generation
 export { generateTailwindConfig } from './tailwind.config';